+++ test/mocks/sync_config_test_mocks.dart
@@ -1,5 +1,3 @@
-import 'package:lotti/blocs/sync/outbox_cubit.dart';
-import 'package:lotti/blocs/sync/outbox_state.dart';
 import 'package:lotti/database/sync_db.dart';
 import 'package:mocktail/mocktail.dart';
 
@@ -11,21 +9,6 @@ MockSyncDatabase mockSyncDatabaseWithCount(int count) {
 
   when(mock.watchOutboxCount).thenAnswer(
     (_) => Stream<int>.fromIterable([count]),
-  );
-
-  return mock;
-}
-
-class MockOutboxCubit extends Mock implements OutboxCubit {}
-
-MockOutboxCubit mockOutboxCubit(OutboxState outboxState) {
-  final mock = MockOutboxCubit();
-  when(() => mock.state).thenReturn(outboxState);
-
-  when(mock.close).thenAnswer((_) async {});
-
-  when(() => mock.stream).thenAnswer(
-    (_) => Stream<OutboxState>.fromIterable([outboxState]),
   );
 
   return mock;